Rename file to match the export
