+++ src/components/EmojiPicker/EmojiPicker.js
